Upgrade versions of plugins and Netbeans target
Many of the versions of various plugins and dependencies are outdated so
they have been upgraded to the most recent compatible release versions.
The plugins modified are as follows:

    - nmb-maven-plugin 3.13 => 4.1
    - maven-compiler-plugin 3.1 => 3.6.0
    - maven-jar-plugin 2.4 => 3.0.2
    - maven-release-plugin 2.5 => 2.5.3
    - maven-source-plugin 2.2.1 => 3.0.1
    - maven-javadoc-plugin 2.9.1 => 2.10.4
    - maven-deploy-plugin 2.8.1 => 2.8.2
    - versions-maven-plugin 2.1 => 2.3
    - maven-scm-plugin 1.9 => 1.9.5

The upgrade of the maven-jar-plugin necessitated a minor change to the
plugin's configuration. The "<useDefaultManifestFile>" option had been
removed as of the 3.x.x releases in favor of the standardized "archiver"
configuration element. The modification was simply to replace the
removed option with an archiver formatted configuration which achieves
exactly the same functionality. Instead of implicitly including the
default Manifest it is now explicitly included by declaring the default
path of the Manifest (which is /META-INF/MANIFEST.MF).

Another minor change was needed in the POM as a result of upgrading the
versions of the plugins, namely a declaration of a prerequisite of Maven
version of 3.0.5 or later. The version 3.0.5 was used as the
prerequisite because that is minimum version required for the declared
plugins.

The dependencies of the project were upgraded, including:

    - commons-io 2.4 => 2.5
    - antlr4-runtime 4.5 => 4.5.3
    - <netbeans.api.version> RELEASE80 => RELEASE802

The upgrade to the Netbeans version is still compatible with IDE version
8.0, though any additional upgrade would require breaking that backwards
compatibility.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>