Fix Java 7 compatibility and modify CI config
The project README indicates that the plugin needs a runtime environment
of at minimum Java 8, however the compiler was generating Java 7
compatible code. The README has been changed to reflect the actual
required Java version, which is Java 7 or later. That said however, the
compiler simply verifies that the code compiles with your local JDK and
that the code doesn't use any illegal keywords, etc. for the source or
target versions. This problem means it is possible to successfully
compile a "Java 7" JAR that contains an illegal call to a Java 8+ API.
Even worse, if a malformed JAR is distributed, the error only surfaces
at runtime, likely when it is used by the end-user. To avoid this
problem the build now executes the animal-sniffer plugin during
compilation. This plugin verifies that the code is API compatible with
the Java 7 runtime API and generates an error if the verification fails.
Additionally, during compilation the compiler is now configured to emit
all warnings and to attempt to optimize the generated Java classes.
After enabling warnings a method annotation was used in order to silence
an unchecked cast warning within auto-generated UI code of the
EditorConfigOptionPanel class.

Future Travis-CI builds should notice a significant speedup due to the
addition of a caching declaration to the configuration file. Travis-CI,
when possible, will cache the Maven repository contents downloaded
during the course of building the project. This allows CI builds to
prevent redundantly downloading the same artifacts during every build
and job. Normally such caching might present a reproducibility problem,
but the Maven repository artifacts are hash verified, making it safe to
cache the data. A minor additional change to the .travis.yml file is the
addition of a YAML version directive. The YAML version declared was YAML
1.1, as that is the highest Travis-CI supports, though they may support
YAML 1.2 in the future.

The generated NBM artifact now includes the author and homePageURL
information which was previously missing. Now that this information is
added the plugin can potentially be distributed via the Netbeans Plugin
Center.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>