Fix 2 for pull request 101 review
Remove the warning suppression from the generated UI code of the
EditorConfigOptionsPanel class and fix the unchecked cast by explicitly
adding the needed String type declaration. Also fix an error overlooked
in the review regarding the caching of the Maven artifacts downloaded
during the Travis-CI build process. For build determinism and
reproducibility the downloaded artifacts must have the hash checksum
verified, which previously was not explicitly required. Now, through the
addition of the maven.config file containing the fail-fast and
strict-checksums options, Maven correctly verifies the hashes of all
artifacts and fails immediately should such a hash checksum verification
fail.

These changes are some of the changes mentioned during the pull request
review, which can be viewed here:

https://github.com/welovecoding/editorconfig-netbeans/pull/101

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>